moved componets to Shared and gave it the right dependancies to work
