+++ Content.Shared/Forensics/Components/CleansForensicsComponent.cs
@@ -1,14 +1,16 @@
-namespace Content.Server.Forensics;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Forensics.Components;
 
 /// <summary>
 /// This component is for items that can clean up forensic evidence
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
 public sealed partial class CleansForensicsComponent : Component
 {
     /// <summary>
     /// How long it takes to wipe prints/blood/etc. off of things using this entity
     /// </summary>
-    [DataField]
+    [DataField, AutoNetworkedField]
     public float CleanDelay = 12.0f;
 }
