+++ Content.Shared/Forensics/Components/DnaSubstanceTraceComponent.cs
@@ -1,9 +1,11 @@
-namespace Content.Server.Forensics;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Forensics.Components;
 
 /// <summary>
 /// This component stops the entity from leaving finger prints,
 /// usually so fibres can be left instead.
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent]
 public sealed partial class DnaSubstanceTraceComponent : Component
 { }
