+++ Content.Shared/Forensics/Components/FiberComponent.cs
@@ -1,10 +1,12 @@
-namespace Content.Server.Forensics
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Forensics.Components
 {
     /// <summary>
     /// This controls fibers left by gloves on items,
     /// which the forensics system uses.
     /// </summary>
-    [RegisterComponent]
+    [RegisterComponent, NetworkedComponent]
     public sealed partial class FiberComponent : Component
     {
         [DataField]
