+++ Content.Shared/Forensics/Components/ForensicPadComponent.cs
@@ -1,12 +1,14 @@
-namespace Content.Server.Forensics
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Forensics.Components
 {
     /// <summary>
     /// Used to take a sample of someone's fingerprints.
     /// </summary>
-    [RegisterComponent]
+    [RegisterComponent, NetworkedComponent, AutoGenerateComponentState]
     public sealed partial class ForensicPadComponent : Component
     {
-        [DataField("scanDelay")]
+        [DataField("scanDelay"), AutoNetworkedField]
         public float ScanDelay = 3.0f;
 
         public bool Used = false;
