+++ Content.Shared/Forensics/Components/IgnoresFingerprintsComponent.cs
@@ -1,7 +1,9 @@
-namespace Content.Server.Forensics.Components;
+using Robust.Shared.GameStates;
+
+namespace Content.Shared.Forensics.Components;
 
 /// <summary>
 /// This component is for entities we do not wish to track fingerprints/fibers, like puddles
 /// </summary>
-[RegisterComponent]
+[RegisterComponent, NetworkedComponent]
 public sealed partial class IgnoresFingerprintsComponent : Component { }
